requirements.txt
Django==2.0.4
requests==2.18.4
django-model-utils==3.1.1
djangorestframework==3.8.2
ipython==6.3.0
pytest==3.5.1
pytest-django==3.1.2
